+++ lib/components/card/gf_card.dart
@@ -176,7 +176,9 @@ class GFCard extends StatelessWidget {
                   _defaultClipBehavior,
               child: imageOverlay == null ? cardChild : overlayImage,
             )
-          : imageOverlay == null ? cardChild : overlayImage,
+          : imageOverlay == null
+              ? cardChild
+              : overlayImage,
     );
   }
 }