+++ lib/components/dropdown/gf_dropdown.dart
@@ -1,7 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:getwidget/getwidget.dart';
 
 class GFDropdown<T> extends StatefulWidget {
-  GFDropdown(
+  /// GF Dropdown let user to select from the number of items and display selected
+  /// item in the button. It displays list of items in the overlay dropdown fashion.
+  const GFDropdown(
       {Key key,
       @required this.items,
       this.icon,
@@ -17,22 +20,24 @@ class GFDropdown<T> extends StatefulWidget {
       this.iconDisabledColor,
       this.iconEnabledColor,
       this.iconSize = 24.0,
-      this.isDense = false,
+      this.isDense = true,
       this.isExpanded = false,
-      this.itemHeight = kMinInteractiveDimension,
+      this.itemHeight = 40,
       this.focusColor,
       this.focusNode,
       this.autofocus = false,
       this.dropdownColor,
       this.padding = const EdgeInsets.all(5),
       this.borderRadius = const BorderRadius.all(Radius.circular(4)),
-      this.borderColor});
+      this.border = const BorderSide(
+          color: Colors.transparent, width: 1, style: BorderStyle.solid),
+      this.dropdownButtonColor = GFColors.WHITE})
+      : super(key: key);
 
-  DropdownButtonBuilder selectedItemBuilder;
+  final DropdownButtonBuilder selectedItemBuilder;
   final List<DropdownMenuItem<T>> items;
 
   /// The widget to use for the drop-down button's icon.
-  ///
   /// Defaults to an [Icon] with the [Icons.arrow_drop_down] glyph.
   final Widget icon;
 
@@ -42,10 +47,10 @@ class GFDropdown<T> extends StatefulWidget {
   /// The value of the currently selected [DropdownMenuItem].
   final T value;
 
-  /// the color of the border of the dropdown button
-  final Color borderColor;
+  /// Defines the border of dropdown button
+  final BorderSide border;
 
-  ///The padding given inside the dropdown
+  /// Defines the padding given inside the dropdown
   final EdgeInsets padding;
 
   /// A placeholder widget that is displayed by the dropdown button.
@@ -55,7 +60,6 @@ class GFDropdown<T> extends StatefulWidget {
   final Widget disabledHint;
 
   /// Called when the user selects an item.
-  ///
   /// If the [onChanged] callback is null or the list of [DropdownButton.items]
   /// is null then the dropdown button will be disabled,
   final ValueChanged<T> onChanged;
@@ -87,26 +91,27 @@ class GFDropdown<T> extends StatefulWidget {
   /// Set the dropdown's inner contents to horizontally fill its parent.
   final bool isExpanded;
 
-  /// The default value is [kMinInteractiveDimension], which is also the minimum
-  /// height for menu items.
-  ///
-  /// If this value is null and there isn't enough vertical room for the menu,
+  /// Defines the height of the menu items
   final double itemHeight;
 
   /// The color for the button's [Material] when it has the input focus.
   final Color focusColor;
 
-  /// {@macro flutter.widgets.Focus.focusNode}
+  /// Defines the keyboard focus for this widget.
   final FocusNode focusNode;
 
-  /// {@macro flutter.widgets.Focus.autofocus}
+  /// On true state it should focus itself if nothing else is already focused.
+  /// Defaults to false
   final bool autofocus;
 
-  /// The background color of the dropdown.
+  /// Defines the background color of the dropdown.
   final Color dropdownColor;
 
-  /// The border radius  of the dropdown.
+  /// Defines the border radius  of the dropdown.
   final BorderRadius borderRadius;
+
+  /// Defines the background color of the dropdownButton.
+  final dynamic dropdownButtonColor;
 
   @override
   _GFDropdownState createState() => _GFDropdownState();
@@ -114,38 +119,37 @@ class GFDropdown<T> extends StatefulWidget {
 
 class _GFDropdownState extends State<GFDropdown> {
   @override
-  Widget build(BuildContext context) {
-    return Card(
+  Widget build(BuildContext context) => Material(
+        color: widget.dropdownButtonColor,
+        shape: RoundedRectangleBorder(
+          side: widget.border,
+          borderRadius: widget.borderRadius,
+        ),
         child: Container(
-            padding: widget.padding,
-            decoration: BoxDecoration(
-                borderRadius: widget.borderRadius,
-                border: Border.all(
-                    color: widget.borderColor != null
-                        ? widget.borderColor
-                        : Colors.white)),
-            child: DropdownButton(
-            
-              items: widget.items,
-              selectedItemBuilder: widget.selectedItemBuilder,
-              value: widget.value,
-              hint: widget.hint,
-              disabledHint: widget.disabledHint,
-              onChanged: widget.onChanged == null ? null : widget.onChanged,
-              onTap: widget.onTap,
-              elevation: widget.elevation,
-              style: widget.style,
-              icon: widget.icon,
-              iconDisabledColor: widget.iconDisabledColor,
-              iconEnabledColor: widget.iconEnabledColor,
-              iconSize: widget.iconSize,
-              isDense: widget.isDense,
-              isExpanded: widget.isExpanded,
-              itemHeight: widget.itemHeight,
-              focusColor: widget.focusColor,
-              focusNode: widget.focusNode,
-              autofocus: widget.autofocus,
-              dropdownColor: widget.dropdownColor,
-            )));
-  }
+          height: widget.itemHeight,
+          padding: widget.padding,
+          child: DropdownButton(
+            items: widget.items,
+            selectedItemBuilder: widget.selectedItemBuilder,
+            value: widget.value,
+            hint: widget.hint,
+            disabledHint: widget.disabledHint,
+            onChanged: widget.onChanged,
+            onTap: widget.onTap,
+            elevation: widget.elevation,
+            style: widget.style,
+            icon: widget.icon,
+            iconDisabledColor: widget.iconDisabledColor,
+            iconEnabledColor: widget.iconEnabledColor,
+            iconSize: widget.iconSize,
+            isDense: widget.isDense,
+            isExpanded: widget.isExpanded,
+            // itemHeight: widget.itemHeight,
+            focusColor: widget.focusColor,
+            focusNode: widget.focusNode,
+            autofocus: widget.autofocus,
+            dropdownColor: widget.dropdownColor,
+          ),
+        ),
+      );
 }